Added new variables to dummy module for library use

+++ src/mc.f90
@@ -40,9 +40,9 @@ module mc
   !---------------------------------------------------------------------------!
   !                      P r i v a t e   V a r i a b l e s                    !
   !---------------------------------------------------------------------------!
-  real(kind=dp) :: mc_target_ratio = 0.5_dp     ! Target acceptance ratio
-  logical       :: eq_adjust_mc    = .true.     ! Do we adjust MC to reach it
-  integer       :: eq_adjust_int   = 100        ! Interval at which to adjust
+  real(kind=dp)     :: mc_target_ratio = 0.5_dp     ! Target acceptance ratio
+  logical           :: eq_adjust_mc    = .true.     ! Do we adjust MC to reach it
+  integer(kind=it)  :: eq_adjust_int   = 100        ! Interval at which to adjust
 
   integer(kind=it) :: eq_mc_cycles  = 10000      ! Equilibration cycles 
   integer(kind=it) :: max_mc_cycles = 500000000  ! How many cycles to perform
